ltoma: off_t fix lost in last update
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@37564 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ include/iostream/ami_sort.h
@@ -30,7 +30,7 @@
  *  This program is distributed in the hope that it will be useful,
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- *  General Public License for more details.  *
+ *  General Public License for more details.
  *  **************************************************************************/
 
 
@@ -97,7 +97,7 @@ AMI_sort(AMI_STREAM<T> *instream, AMI_STREAM<T> **outstream, Compare *cmp,
 {
   char* name=NULL;
   queue<char*>* runList;
-  int instreamLength;
+  off_t instreamLength;
 
   assert(instream && outstream && cmp); 
   instreamLength = instream->stream_len();
